anti add bot and react role/requirements.txt
discord.py
python-dotenv
requests
#yung procfile para lang yan sa railway para ma run mo kahit i kill mo yung terminal mo
